Refactor XP commands and make them more discreet
