+++ configuration.json
@@ -260,5 +260,6 @@
     "ohno": "<:ohno:930365904657723402>",
     "tendi_smile_happy": "<:tendi_smile_happy:757768236069814283>",
     "tgg_love_heart": "<:love_heart_tgg:798600300264161280>"
-  }
+  },
+  "all_emoji": []
 }
